requirements.txt
asgiref==3.5.2
boto3==1.24.1
botocore==1.27.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.5
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.1.0
